requirements.txt
tqdm==4.30.0
torchvision==0.2.1
imgaug==0.2.7
torch==1.0.0
matplotlib==3.0.2
numpy==1.16.0
Pillow==5.4.1
